Working on Android (very hard)
It's very hard to make the Android version work, due to the very strict security policy on executing binaries, also the lack of open-source information, will write a tutorial about it. Also found that linux binary for x86-64 will not run on x86-64 emulator, while when debugging on physical android device, Visual Studio failed to attach the process to it, luckily, the arm64 binary seems to work, but the file path on android from file picker seems to be broken, and no way to find about it... very difficult.